+++ src/registers/vcellsum.rs
@@ -7,7 +7,7 @@ pub struct VCellSum(u16);
 impl VCellSum {
     /// Get the sum of cell voltages measurement code
     pub const fn get_vcellsum_meas(&self) -> u16 {
-        (self.0 & 0x7FFF) as u16
+        self.0 & 0x7FFF
     }
 }
